+++ src/main/java/leaf/soulhome/constants/Constants.java
@@ -166,6 +166,8 @@ public class Constants
         public static final String ABILITY_CLEANSING_FONT_CLEANSED = "message.soulhome.ability.cleansing_font.cleansed";
         public static final String ABILITY_ENCORE_ALONE = "message.soulhome.ability.encore.alone";
         public static final String ABILITY_ENCORE_SHARED = "message.soulhome.ability.encore.shared";
+        public static final String ABILITY_LAST_STAND_UNHURT = "message.soulhome.ability.last_stand.unhurt";
+        public static final String ABILITY_CALMING_SMOKE_NOTHING = "message.soulhome.ability.calming_smoke.nothing";
 
         // Two refusals every active shares. A press that does nothing and says nothing is
         // indistinguishable from a mod that is broken, which is how a recharging Thunderclap and
@@ -186,7 +188,8 @@ public class Constants
         public static final String ABILITY_NAME_CLEANSING_FONT = "ability.soulhome.cleansing_font";
         public static final String ABILITY_NAME_UPDRAFT = "ability.soulhome.updraft";
         public static final String ABILITY_NAME_ENCORE = "ability.soulhome.encore";
-        public static final String ABILITY_NAME_REQUIEM = "ability.soulhome.requiem";
+        public static final String ABILITY_NAME_LAST_STAND = "ability.soulhome.last_stand";
+        public static final String ABILITY_NAME_CALMING_SMOKE = "ability.soulhome.calming_smoke";
     }
 
     public static class NBTKeys